Move admin mapping utils to admin.v1.common module
